CAMEL-10691: HttpRestServletResolveConsumerStrategy should pick the path with longest prefix match
